Updated reddit collector to work with correct mappings

+++ ProblemCrawler.Collectors/ProblemCrawler.Collectors.Reddit/Extensions/RedditCollectorMappingProfile.cs
@@ -13,7 +13,7 @@ public class RedditCollectorMappingProfile : Profile
             .ForMember(dest => dest.ItemType, opt => opt.MapFrom(_ => "Post"))
             .ForMember(dest => dest.Source, opt => opt.MapFrom(_ => "Reddit"))
             .ForMember(dest => dest.Content, opt => opt.MapFrom(src => src.IsSelf ? src.Selftext ?? string.Empty : src.Url ?? string.Empty))
-            .ForMember(dest => dest.CreatedAt, opt => opt.MapFrom(src => DateTimeOffset.FromUnixTimeSeconds(src.CreatedUtc).UtcDateTime))
+            .ForMember(dest => dest.CreatedAt, opt => opt.MapFrom(src => DateTimeOffset.FromUnixTimeSeconds((long)src.CreatedUtc).UtcDateTime))
             .ForMember(dest => dest.SourceUrl, opt => opt.MapFrom(src => src.Permalink != null ? $"{RedditCollectorDefaults.BaseUrl}{src.Permalink}" : null))
             .ForMember(dest => dest.Metadata, opt => opt.Ignore())
             .AfterMap((src, dest) => dest.Metadata = BuildPostMetadata(src));
@@ -22,7 +22,7 @@ public class RedditCollectorMappingProfile : Profile
             .ForMember(dest => dest.ItemType, opt => opt.MapFrom(_ => "Comment"))
             .ForMember(dest => dest.Source, opt => opt.MapFrom(_ => "Reddit"))
             .ForMember(dest => dest.Content, opt => opt.MapFrom(src => src.Body))
-            .ForMember(dest => dest.CreatedAt, opt => opt.MapFrom(src => DateTimeOffset.FromUnixTimeSeconds(src.CreatedUtc).UtcDateTime))
+            .ForMember(dest => dest.CreatedAt, opt => opt.MapFrom(src => DateTimeOffset.FromUnixTimeSeconds((long)src.CreatedUtc).UtcDateTime))
             .ForMember(dest => dest.SourceUrl, opt => opt.MapFrom(src => src.Permalink != null ? $"{RedditCollectorDefaults.BaseUrl}{src.Permalink}" : null))
             .ForMember(dest => dest.Metadata, opt => opt.Ignore())
             .AfterMap((src, dest) => dest.Metadata = BuildCommentMetadata(src));
